Security

+++ crud/commands/delete.go
@@ -14,34 +14,39 @@ func Delete(w http.ResponseWriter, r *http.Request) {
 	utils.SetCorsHeaders(&w)
 
 	if r.Method == "DELETE" {
-		//connect the database
-		db := utils.DbConnect()
 
-		//get url values
-		vars := mux.Vars(r)
-		id_command := vars["id_command"]
+		if utils.Securized(w, r) {
+			//connect the database
+			db := utils.DbConnect()
 
-		//delete command lines with id_command linked
-		sqlQuery := ("DELETE FROM command_line WHERE id_command = " + id_command + ";")
+			//get url values
+			vars := mux.Vars(r)
+			id_command := vars["id_command"]
 
-		//execute the sql query
-		_, err := db.Exec(sqlQuery)
-		utils.CheckErr(err)
+			//delete command lines with id_command linked
+			sqlQuery := ("DELETE FROM command_line WHERE id_command = " + id_command + ";")
 
-		//create the sql query
-		sqlQuery = ("DELETE FROM command WHERE id = " + id_command + ";")
+			//execute the sql query
+			_, err := db.Exec(sqlQuery)
+			utils.CheckErr(err)
 
-		//execute the sql query
-		_, err = db.Exec(sqlQuery)
-		utils.CheckErr(err)
+			//create the sql query
+			sqlQuery = ("DELETE FROM command WHERE id = " + id_command + ";")
 
-		//close the database connection
-		db.Close()
+			//execute the sql query
+			_, err = db.Exec(sqlQuery)
+			utils.CheckErr(err)
 
-		//create the json response
-		json.NewEncoder(w).Encode(structures.JsonResponseCommand{
-			Type:    "success",
-			Message: "Command deleted",
-		})
+			//close the database connection
+			db.Close()
+
+			//create the json response
+			json.NewEncoder(w).Encode(structures.JsonResponseCommand{
+				Type:    "success",
+				Message: "Command deleted",
+			})
+		} else {
+			w.WriteHeader(http.StatusUnauthorized)
+		}
 	}
 }
